requirements.txt
# Browser Automation
playwright==1.40.0

# HTML Parsing / DOM Analysis
beautifulsoup4==4.12.2
lxml==4.9.3

# LLM Integration
openai==1.6.1
google-generativeai==0.3.2

# Backend API
fastapi==0.108.0
uvicorn==0.25.0

# UI
streamlit==1.29.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.2
